% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.STS.R
\name{distrib.STS}
\alias{distrib.STS}
\title{Empirical Distribution Function}
\usage{
distrib.STS(Xrand, x = Xrand)
}
\arguments{
\item{Xrand}{is a vector with observations of a random variable}

\item{x}{is a vector which will be evaluate the Empirical Distribution Function based on \code{Xrand}}
}
\value{
a vector which contain the values of Empirical Distribution Function evaluated in \code{x}
}
\description{
This function returns the Empirical Distribution Function based on a vector containig observations of a random variable
}
