% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.STS.R
\name{impute.STS}
\alias{impute.STS}
\title{Imputation of Seasonal Time Series}
\usage{
impute.STS(TS, seasonality = 12, trend = NULL, ...)
}
\arguments{
\item{TS}{is a object of class \code{ts}, time series (either \code{ncol()>=2})}

\item{seasonality}{either the character string "periodic" or the span (in lags) of the loess window for seasonal extraction}

\item{trend}{he span (in lags) of the loess window for trend extraction, which should be odd. If \code{NULL}, the default, \code{nextodd(ceiling((1.5*period) / (1-(1.5/s.window))))}, is taken.}

\item{...}{Extra arguments passed to \code{stlplus} function from \code{stlplus} package from "https://github.com/hafen/stlplus"}
}
\value{
returns an object of class "ts", containing the reconstructed time series
}
\description{
This function aproximate by interpolation the values of a time series (unknonw), based in the values of near (geographically) time series.
}
\examples{
# Univariate Time Serie imputation
x <- co2
x[sample(1:length(x),30,FALSE)] <- NA
sts <- impute.STS(x)

# Multivariate Time Serie imputation
z <- rnorm(300)
z[sample(1:length(z),60,FALSE)] <- NA
z <- sin(1:length(z)) + z
z <- matrix(z, 100, 3)
z <- ts(z, start = c(1961, 1), frequency = 12)
sts <- impute.STS(z)
}
