% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSTS.R
\docType{package}
\name{imputeSTS}
\alias{imputeSTS}
\title{imputeSTS}
\description{
Contains the implementation of methodology created with the objective of
the imputation of missing data in seasonal time series based on STL-Loess Descomposition
(Local Non Paramétric Regression) and Monte Carlo simulation. The package also has
functions for visualization.
}
\author{
Cristian Pachacama \email{cristian.pachacama01@gmail.com}
}
