% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.STS.R
\name{plot.STS}
\alias{plot.STS}
\title{Plot of Cleaned and original Time Series}
\usage{
\method{plot}{STS}(STSclean, k = NULL)
}
\arguments{
\item{STSclean}{is a object type \code{STS} and It contain imputation of TS time series, by default is \code{NULL}, in that case \code{TSclean = imputeSTS(TS)}}

\item{k}{is the number of column (time serie) from \code{STSclean} to be plot, by default is \code{NULL} in that case all time series in TS are ploting}
}
\value{
a plot with the original series and the reconstructed series are found
}
\description{
This function generates a plot with the original and reconstructed series
}
\examples{
x <- co2
x[sample(1:length(x),20,FALSE)] <- NA
sts <- impute.STS(x)
plot(sts)
}
